+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   - Example: 10.2.1.4 is the 5th version that supports khiops 10.2.1.
 - Internals: Changes in *Internals* sections are unlikely to be of interest for data scientists.
 
+## Unreleased
+
+### Added
+- (`sklearn`) Support for boolean and float targets in `KhiopsClassifier`.
+
 ## 10.2.4.0 - 2024-12-19
 
 ### Added